fps added to Breakout and Asteroids added to Games

+++ Breakout/src/Update.h
@@ -1,4 +1,4 @@
 #pragma once
-void update(SDL_Renderer* renderer, int time);
+void update(SDL_Renderer* renderer, float time);
 bool checkRectCollisionBricks(SDL_Rect rect, SDL_Renderer* renderer, bool ball, int ballIndex);
 void nextLevel(SDL_Renderer* renderer);
